rbd: return image id in info command output
And as a side effect, previously when rbd info was called
with --image-id, the image id was shown instead of the image name,
so the user was not able to see the image name from info output.
Now it will show the image name.

Signed-off-by: Mykola Golub <mgolub@suse.com>
(cherry picked from commit 4784d6202bc670e3c74a426f5f7acda2eedd8fe7)

Conflicts:
	src/test/cli-integration/rbd/formatted-output.t
	src/tools/rbd/action/Info.cc