mon: use min-delay propose to gather pg_temp and up_thru updates
The above two messages are essential for the peering process,
which will prevent a PG from serving write and read operations
until it completes.
With this patch mon shall now be able to respond a pg_temp/up_thru
changing request faster and hence a reduced chance of blocking
client I/Os.

Signed-off-by: xie xingguo <xie.xingguo@zte.com.cn>