os/bluestore: fix allocation failure despite of enough space
Bluestore normally use a standard small alloc unit (e.g., 16kB).
Hence the offset of a free segment from "slow" probably can not
meet the 1MB alignment that bluefs claim to support.

In case we're running low on space on "slow" device or
the space tree just get too fragmented, we force to use a
minimal alignment (4kB) when we have to switch to using
best-fit allocations.

Signed-off-by: xie xingguo <xie.xingguo@zte.com.cn>