common/options: 16kB bluefs_alloc_size
It turns out that we have workloads that need to handle
tons of small objects, as a result of which the space
of "slow" can become extremely fragmented and make it's hard
to transmit free space between bluestore and bluefs.

Hard cap bluefs_alloc_size at 16kB to keep pace with bluestore.
Hence the balancing procedure shall always succeed as long as
there is enough space there.

Signed-off-by: xie xingguo <xie.xingguo@zte.com.cn>