osd: prime upmap for specific osds
With this feature on, we now automatically generate pg_upmaps for
specific osds, e.g., osds that are newly added into the cluster,
or removed from the cluster. We can leave it to balancer to
(slowly, presumably) cancel these pg_upmaps to approach the final
pg distribution without causing any sensable performance issues to
on-going client activities.
Note that require_min_compat_client >= luminous is still mandatory,
which means older clients won't be able to enjoy the benefit.

Signed-off-by: xie xingguo <xie.xingguo@zte.com.cn>