mon/OSDMonitor: enable pg log hard limit for new clusters
The feature is useful for both we not running out of memory under
certain circumstances and consistently better performance.
The reason this wasn't included in the luminous backport was for
compatibility with existing mimic releases. That is, there were mimic
releases that did not support this, so if someone upgraded from luminous
to those releases, the hardlimit would not be effective.
It's added to new clusters in nautilus and later.

Note that we could still enable this by manually running the
"ceph osd set pglog_hardlimit" command for new 12.2.13 clusters,
but it'd be simpler to default it on since we have decided to skip
the entire mimic release!

Signed-off-by: xie xingguo <xie.xingguo@zte.com.cn>